refactor(index): Update imports in EditorGroups.kt and SwitchRequest.kt
Updated imports in EditorGroups.kt to use IndexCache from the index package.
Added a SwitchRequest data class in SwitchRequest.kt to represent file switch requests.
