+++ src/main/java/krasa/editorGroups/index/IndexNotReady.kt
@@ -1,3 +1,3 @@
-package krasa.editorGroups
+package krasa.editorGroups.index
 
 class IndexNotReady(s: String?, e: Exception?) : Exception(s, e)
